Extract PortForwarder tests to own suite
